contracts: type Contract.right as Option<OptionRight> (typed-status sweep PR 2)
Promote Contract.right from String to Option<OptionRight>. The existing
OptionRight enum gains #[non_exhaustive], FromStr<Err = Error>, ToField,
Hash, Serialize, Deserialize. Empty wire is None; unknown wire surfaces
as Error::Parse instead of a raw string (rule 16).

Contract::option's 4th param changes from &str to OptionRight.
ContractBuilder::right takes OptionRight directly; the runtime
"P/C" case-insensitive validation is deleted (structurally impossible
once typed). Two unreachable tests removed.

FromStr accepts only canonical "C"/"P" - verified no "CALL"/"PUT" wire
in any captured-bytes fixture or C# reference; the long-form note was
doc cruft.

Modernize touched module (rule 9): the inline #[cfg(test)] mod tests
block at the bottom of src/contracts/mod.rs (~220 lines) lifted to a
sibling mod_tests.rs.

Side fix: testdata/builders/orders.rs had right: "?" - that is a
sample-app display fallback for empty string, not a TWS wire value
(verified against tws-api/samples). Changed to "" so the new strict
decoder accepts it.
