+++ examples/async/calculate_option_price.rs
@@ -19,7 +19,10 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     println!("=== Calculating Option Price ===");
     println!(
         "Contract: {} {} {} @ {}",
-        contract.symbol, contract.last_trade_date_or_contract_month, contract.right, contract.strike
+        contract.symbol,
+        contract.last_trade_date_or_contract_month,
+        contract.right.map_or("", |r| r.as_str()),
+        contract.strike
     );
     println!("Volatility: {:.1}%", volatility * 100.0);
     println!("Underlying Price: ${underlying_price:.2}");
